==> scripts/both/00_libs.R <==
library(here)
library(fs)
library(tidyverse)
library(ggpubr)
library(brms)
library(xtable)
library(sjPlot)
library(insight)
library(httr)
library(ggrepel)
library(here)
library(tidyverse)
library(stringr)
library(scales)
library(ggdark)
library(ggrepel)
library(ggpubr)
library(brms)
library(tidybayes)
library(modelr)
library(peRspective)
library(bayestestR)
